Fix for #97
For some reason I don't understand the call to `Her::Model.included`
does not seem to properly set the default_api when inheriting from a
base class in a module. I work around this by explicitly calling
Her::API.default_api if `.use_api` returns nil or the superclass does
not respond to `:use_api`

This seems gross, but it enables me to use 0.6.x with my app - without
this fix anything above 0.5.5 would break my app/tests.

Note that this was a bit of a hisenbug, and attempts to use `spawn_class`
and even `subject` in the specs meant that the results would change, so
the spec for this is explicit as possible.