Save a reference to fetched has one relationships.
Currently has_one relationships are fetched everytime a message is sent
to the proxy. This commit adds a has_one cache that prevents the proxy
from fetching a new record if it has already fetched model with the
current params.

I added specs that show the behavior I am trying to accomplish. I think
the specs are straight forward, but if you run the against master they
will fail.