Memoize user-defined methods during Her attribute definition.
585e91c fixed a bug where attributes defined by Her would overwrite
user-defined methods, however, it also appears to have introduced a
performance regression. According to the profiler tool ruby-prof, when
initializing new objects, `Module#instance_methods` takes up a
significant chunk of time (observed with Ruby 1.9.3 and 2.1.5).

If it's safe to assume that it's unlikely the set of user-defined
methods will change *while* this method is being evaluated, memoizing
them can result in a performance boost.