Allow customizable JSON serialization mirroring ActiveModel's
One use case I have for the Her gem is to retrieve resources from a
remote app to be used in the Javascript client of the local app.
This entails serializing the Her models into JSON to be returned by
a controller in the local app.

This commit introduces `#as_json` on both the `Model` and `Relation`
so that when `#to_json` is invoked, it will return a nicely formatted
JSON string that a Javascript (or other) client of the local app can
easily consume.

The interface was modeled after `ActiveModel::Serializers::JSON` and
accepts the `:only`, `:except`, and `:methods` options. This is an
incomplete copy of `ActiveModel` functionality, as it does not accept
the `:root` or `:include` options. However, it satisfies my current
use case and can be extended to accept those options in the future
as other developers need them.