Handle the destroyed instance variable accordingly to the presence of response_errors
The presence of response_errors for a destroy request indicates the
request has not been successful, e.g. due to authorisation issues. By
returning a truthy value a false indication is given as we cannot
assume everyone checking for response errors.

I therefore relate the value of @destroyed to the existence of errors
as it seems more sensible to me.

Looking forward to your comments.