Fix embeded_params method for all associations
The embeded_params method was fixed to properly handle all associations
including the has_one, has_many, and belongs_to. In addition, the
class_name was required to fix two specs. This problem in the specs
did not previously surface because the has_one and belongs_to
associations were being ignored.