Update the setter methods cache after defining new attributes
Not doing so resulted in very strange behaviour. Any previously unseen
attributes passed to new with send_only_modified_attributes enabled
would be missing from the request parameters. This is because change
tracking is only effective for setter methods.

Now this is fixed, it is unclear to me what the @attributes.merge!
line in assign_attributes is for. In theory, all the given attributes
(that are not associations) should have setter methods defined by this
point. It should probably still trigger change tracking in any case.