Drop the setter methods cache in favour of respond_to? (see #472)
The cache was not being updated when new attributes were added,
causing very strange behaviour. Any previously unseen attributes
passed to new with send_only_modified_attributes enabled would be
missing from the request parameters. This is because change tracking
is only effective for setter methods.

Updating the cache turned out to be more expensive than not having a
cache at all. Using respond_to? would be fastest but this breaks
things as respond_to_missing? returns true for any assignment method.

If method(:foo=).source_location returns nil then this may indicate
that the method is missing but this is also true for methods in native
code.

A faster and more reliable approach is using a fiber-local variable to
make respond_to_missing? return false while doing these checks. It's a
tad ugly but it's the best I can come up with.