Serialize nested Her::Model instances that are not associated
We previously added serialization for ActiveModel::Serialization
instances but Her::Model instances can still be left untouched if they
do not belong to an association.

In order to serialize these safely, we have to move this block of code
into embed_params!, which guards against infinite loops. This has the
added benefit of not doing unnecessary serialization when
send_only_modified_attributes is enabled.